chore: improve feature cards loading
