+++ src/components/FeatureCards/FeatureCards.style.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { Box } from '@mui/material';
+import Box from '@mui/material/Box';
 
 import { styled } from '@mui/material/styles';
