--- src/components/FeatureCards/index.ts
@@ -1,4 +1,0 @@
-export * from '../FeatureCards/FeatureCard';
-export * from '../FeatureCards/FeatureCard.style';
-export * from './FeatureCards';
-export * from './FeatureCards.style';
